Add support for SEPA direct debit payment method
The previous flow for credit cards which uses Stripe's Payment Intents API (1$ card auth which is later voided) does not work with SEPA direct debit, as SEPA direct debit can only be used with an automatic capture. To keep the overall amount of code as low as possible the credit card flow was also adapted such that there is only one flow for both payment methods. Instead of using the Payment Intents API the plugin now uses Setup Intents, i.e. the payment method is only setup for later payments.